I create a new file ServiceType.ts
