+++ client/src/components/ServiceCard/ServiceCard.css
@@ -31,7 +31,6 @@
 }
 
 .service-card .members {
-  /* gap: 16px; */
   padding: 8px;
 }
