+++ client/src/components/ServiceCard/ServiceType.ts
@@ -1,0 +1,11 @@
+export interface MemberProps {
+  id: string;
+  name: string;
+  imgAvatar: string;
+}
+
+export interface ServiceProps {
+  imgCategory: string;
+  category: string;
+  members: MemberProps[];
+}
